% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animaltfdb.180611}
\alias{animaltfdb.180611}
\title{animaltfdb.180611}
\format{tibble}
\usage{
animaltfdb.180611
}
\description{
Tibble version of three tables from the animalTFDB2.0 database.
}
\details{
1. Transcription Factors
2. Chromatin Remodelers
3. Transcriptional Cofactors

1. AnimalTFDB: a comprehensive animal transcription factor database. Hong-Mei Zhang, Hu Chen, Wei Liu, Hui Liu, Jing Gong, Huili Wang and An-Yuan Guo*. Nucl. Acids Res. 2012. 40(Database issue):D144-9.
2. AnimalTFDB 2.0: a resource for expression, prediction and functional study of animal transcription factors. Hong-Mei Zhang$, Teng Liu$, Chun-Jie Liu, Shuangyang Song, Xiantong Zhang, Wei Liu, Haibo Jia, Yu Xue, and An-Yuan Guo*. Nucl. Acids Res. (28 January 2015) 43 (D1): D76-D81.
}
\keyword{datasets}
