% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humantfdb.cofacs.190502}
\alias{humantfdb.cofacs.190502}
\title{humantfdb.tfs.190312}
\format{tibble}
\usage{
humantfdb.cofacs.190502
}
\description{
Tibble version of the TF Cofactor table from the humanTFDB database.
}
\details{
Source:
1. AnimalTFDB: a comprehensive animal transcription factor database. Hong-Mei Zhang, Hu Chen, Wei Liu, Hui Liu, Jing Gong, Huili Wang and An-Yuan Guo*. Nucl. Acids Res. 2012. 40(Database issue):D144-9.

2. AnimalTFDB 2.0: a resource for expression, prediction and functional study of animal transcription factors. Hong-Mei Zhang$, Teng Liu$, Chun-Jie Liu, Shuangyang Song, Xiantong Zhang, Wei Liu, Haibo Jia, Yu Xue, and An-Yuan Guo*. Nucl. Acids Res. (28 January 2015) 43 (D1): D76-D81.

3. 1. AnimalTFDB 3.0: a comprehensive resource for annotation and prediction of animal transcription factors. Hui Hu$, Ya-Ru Miao$, Long-Hao Jia, Qing-Yang Yu, Qiong Zhang and An-Yuan Guo*. Nucl. Acids Res. (11 September 2018).
}
\keyword{datasets}
