% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animaltfdb.180611}
\alias{animaltfdb.180611}
\title{animaltfdb.180611}
\format{tibble}
\usage{
animaltfdb.180611
}
\description{
Tibble version of three tables from the animalTFDB2.0 database.
}
\details{
1. Transcription Factors
2. Chromatin Remodelers
3. Transcriptional Cofactors
}
\keyword{datasets}
